Jsy 찜usermain

+++ src/main/java/com/example/HiMade/admin/controller/AdminStoreController.java
@@ -93,7 +93,7 @@ public class AdminStoreController {
 
     //고정휴무-select
     @GetMapping("/getOffDay")
-    public List<DayOffDay> getOffDay(@RequestParam Integer storeNo){
+    public List<DayOffDay> getOffDay(@RequestParam Long storeNo){
         return adminStoreService.getOffDay(storeNo);
     }
 
@@ -107,7 +107,7 @@ public class AdminStoreController {
 
     //지정휴무 - delete
     @DeleteMapping("/deleteOffSet")
-    public ResponseEntity<Integer> deleteOffSet(@RequestParam Integer storeNo, @RequestParam Integer setNo){
+    public ResponseEntity<Integer> deleteOffSet(@RequestParam Long storeNo, @RequestParam Long setNo){
         Integer result = adminStoreService.deleteOffSet(storeNo, setNo);
         if(result > 0){
             return ResponseEntity.ok(result);
@@ -118,7 +118,7 @@ public class AdminStoreController {
 
     //지정휴무-select
     @GetMapping("/getOffSet")
-    public List<DayOffSet> getOffSet(@RequestParam Integer storeNo){
+    public List<DayOffSet> getOffSet(@RequestParam Long storeNo){
         return adminStoreService.getOffSet(storeNo);
     }
